Move a lot of code around for better structure
